requirements.txt
yt-dlp
paramiko
aiogram
python-dotenv